Update network functions to ESP32 SDK 3.0.x

+++ src/WifiCachingStream.cpp
@@ -99,7 +99,7 @@ void WifiCachingStream::flush()
 
 size_t WifiCachingStream::readBytes(char* buffer, size_t length)
 {
-	int received = 0;
+	size_t received = 0;
 	auto result = network_recv_non_blocking(_connection_sd, buffer, (int)length, &received);
 	if (received >= 1)
 	{
